tegra: dc: Fix cmu LUT1 dt parsing error
Fix LUT1 dt parsing typecast error. While
parsing device tree data, there was a typecast to
u8 in stead of u16 fpr LUT1 values.

Bug 1644175

Reviewed-on: http://git-master/r/745042
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/video/tegra/dc/of_dc.c
@@ -916,7 +916,7 @@ static int parse_cmu_data(struct device_node *np,
 		of_property_for_each_u32(np, "nvidia,cmu-lut1",
 			prop, p, u) {
 			/* OF_DC_LOG("cmu lut1 0x%x\n", u); */
-			*(addr_cmu_lut1++) = (u8)u;
+			*(addr_cmu_lut1++) = (u16)u;
 		}
 	}
